backend/requirements.txt
flask==2.3.0
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.2
tensorflow==2.13.0
opencv-python==4.8.0.74
pillow==10.0.0
numpy==1.24.3
python-dotenv==1.0.0